feature: adding gitflow definition and gitflow image example
Some key benefits include:

	Parallel Development
o	One of the main things about GitFlow is that it allows for parallel development. This is done by isolating new development from finished work and making it easier to distinguish the difference between the two. New development or capability improvements are completed in feature branches and is only merged back into main code when the developer is ready to move to a production state.
	Collaboration
o	Gitflow also offers more collaboration and feature branches which makes it easier for two+ developers to collaborate on the same feature or enhancements. This process makes it easy for other developers to see and follow each collaboration effectively.
	Support For Emergency Fixes
o	GitFlow also supports something called “hotfix branches”. Essentially these hot fix branches can be useds to make an emergency change.

Co-Authored-By: natashakhoblall <natashakhoblall@users.noreply.github.com>