Feature: Added definition for Branch and image example
o	Task: Git branches are effectively a pointer to a snapshot of the changes. When you want to add a new feature or fix a bug you spawn a new branch to encapsulate your changes.
o	Task: added folder for branch which is branch image example
o	Task: added branch photo example

Co-Authored-By: natashakhoblall <natashakhoblall@users.noreply.github.com>